Move fuzz tests to go native fuzz [part1]
