+++ contrib/fuzz/images_fuzz_test.go
@@ -18,6 +18,7 @@ package fuzz
 import (
 	"context"
 	"os"
+	"testing"
 
 	fuzz "github.com/AdaLogics/go-fuzz-headers"
 	"github.com/containerd/containerd/v2/core/images"
@@ -26,21 +27,22 @@ import (
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
-func FuzzImagesCheck(data []byte) int {
-	f := fuzz.NewConsumer(data)
-	desc := ocispec.Descriptor{}
-	err := f.GenerateStruct(&desc)
-	if err != nil {
-		return 0
-	}
-	tmpdir, err := os.MkdirTemp("", "fuzzing-")
-	if err != nil {
-		return 0
-	}
-	cs, err := local.NewStore(tmpdir)
-	if err != nil {
-		return 0
-	}
-	_, _, _, _, _ = images.Check(context.Background(), cs, desc, platforms.Default())
-	return 1
+func FuzzImagesCheck(f *testing.F) {
+	f.Fuzz(func(t *testing.T, data []byte) {
+		f := fuzz.NewConsumer(data)
+		desc := ocispec.Descriptor{}
+		err := f.GenerateStruct(&desc)
+		if err != nil {
+			return
+		}
+		tmpdir, err := os.MkdirTemp("", "fuzzing-")
+		if err != nil {
+			return
+		}
+		cs, err := local.NewStore(tmpdir)
+		if err != nil {
+			return
+		}
+		_, _, _, _, _ = images.Check(context.Background(), cs, desc, platforms.Default())
+	})
 }
